+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - _[OZO-454]_: Added "open in new" icon to signpost, removed hover border in focusable anchor card
 - _[OZO-392]_: Accessibility updates
 - _[OZO-541]_: Clickable datagrid row, border fixes in table and datagrid
+- _[WEB2-6157]_: Extension of the DropZone title parameter
 
 * **Property changes**
   - No changes.